requirements.txt
Jinja2>=3.0
boto3>=1.20
pytest>=7.0
PyYAML>=6.0
